♻️ refactor: simplify auth redirect logic in middleware
- Remove unused buildCallbackUrl and getCallbackUrlOrigin utilities
- Use NEXT_PUBLIC_AUTH_URL directly for single deployment scenarios
- Delete packages/utils/src/server/correctCallbackUrl.* files
- Improve code clarity and reduce runtime overhead



♻️ refactor: merge getSafeOrigin into correctOIDCUrl

- Consolidate getSafeOrigin logic into correctOIDCUrl.ts
- Remove redundant getSafeOrigin.ts file
- Maintain API compatibility through index.ts exports
- Improve code organization and reduce file count



♻️ refactor: inline getSafeOrigin logic into correctOIDCUrl

- Remove standalone getSafeOrigin function export
- Inline origin determination logic directly into correctOIDCUrl
- Simplify code by reducing unnecessary abstraction
- Improve code readability with single cohesive function
