+++ packages/utils/src/server/index.ts
@@ -1,8 +1,5 @@
 export * from './auth';
-export * from './correctCallbackUrl';
 export * from './correctOIDCUrl';
-export * from './geo';
-export * from './getSafeOrigin';
 export * from './response';
 export * from './responsive';
 export * from './sse';
